==> 3. MIND in a soil microcosm/1.SynCom_in_soil_multiomics_customindex_dataprep.Rmd <==
---
title: "MIND for SynCom grown in a soil microcosm: Data Prep"
author: "Oriane Moyne"
date: "10/19/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load libraries

```{r, message = FALSE}
# Packages I commonly use
library(stringr)
library(ggplot2)
library(FactoMineR)
library(factoextra)
library(reshape)
library(reshape2)
library(gplots)
# library(edgeR)
```

## Import data


## Data import (Count table)

```{r}
# Count tables with taxonomy annotations (Woltka gotu output)
syncom <- read.table("/home/moyne/Moyne_disk/Soil_project/Final_datasets/REAL_SOIL_SAMPLES/Expt_2/WoL_subset/Nov_22/custom_index/MIND_Soil_SynCom_Custom_index_Expt2_count_table.tsv", 
                    sep="\t", 
                    dec=".", 
                    header = TRUE, 
                    stringsAsFactors = FALSE, 
                    na.strings = "", 
                    check.names = FALSE, 
                    comment.char = "")

# Correct column names for reproducibility
colnames(syncom) <- str_replace(colnames(syncom), "_custom_index", "")
colnames(syncom) <- str_replace(colnames(syncom), "#FeatureID", "FeatureID")
colnames(syncom) <- str_replace(colnames(syncom), "metaRS_1", "1_metaRS")
colnames(syncom) <- str_replace(colnames(syncom), "metaRS_2", "2_metaRS")
colnames(syncom) <- str_replace(colnames(syncom), "_1_S", "_1_metaG_S")
colnames(syncom) <- str_replace(colnames(syncom), "_2_S", "_2_metaG_S")
colnames(syncom) <- str_replace(colnames(syncom), "_SC_", "_sc_")
colnames(syncom) <- str_replace(colnames(syncom), "_SynCom_", "_syncom_")
colnames(syncom) <- str_split(colnames(syncom), "_S", simplify = TRUE)[,1]
colnames(syncom) <- str_replace(colnames(syncom), "_sc_", "_SynCom_")
colnames(syncom) <- str_replace(colnames(syncom), "_syncom_", "_SynCom_")
colnames(syncom) <- str_replace(colnames(syncom), "Soil_Water", "Soil_Water_T3")
colnames(syncom) <- str_replace(colnames(syncom), "0_1xR2A", "01x")

str_split(colnames(syncom), "_", simplify = TRUE)
```


## Taxonomy Assignment

```{r}
genomes_ID <- str_split(syncom$FeatureID, "_", simplify = TRUE)[,1]
table(genomes_ID) # 18 strains

syncom$genus_strain <- NA
syncom[str_detect(syncom$FeatureID, "MDNCIGFF"), ]$genus_strain <- "Arthrobacter_OAP107"
syncom[str_detect(syncom$FeatureID, "MLEKFOEI"), ]$genus_strain <- "Brevibacillus_OAP136"
syncom[str_detect(syncom$FeatureID, "FOCFADBO"), ]$genus_strain <- "Mucilaginibacter_OAE612"
syncom[str_detect(syncom$FeatureID, "IKHJAPOO"), ]$genus_strain <- "Variovorax_OAS795"
syncom[str_detect(syncom$FeatureID, "LLHGGPMF"), ]$genus_strain <- "Bacillus_OAE603"
syncom[str_detect(syncom$FeatureID, "BKKPEFGL"), ]$genus_strain <- "Bosea_OAE506"
syncom[str_detect(syncom$FeatureID, "FKKIHNNJ"), ]$genus_strain <- "Bradyrhizobium_OAE829"
syncom[str_detect(syncom$FeatureID, "JOPPLEBF"), ]$genus_strain <- "Burkholderia_OAS925"
syncom[str_detect(syncom$FeatureID, "ACPHPGFJ"), ]$genus_strain <- "Chitinophaga_OAE865"
syncom[str_detect(syncom$FeatureID, "JENONDHJ"), ]$genus_strain <- "Niastella_OAS944"
syncom[str_detect(syncom$FeatureID, "PEBMFGFJ"), ]$genus_strain <- "Lysobacter_OAE881"
syncom[str_detect(syncom$FeatureID, "HFAHJMJN"), ]$genus_strain <- "Marmoricola_OAE513"
syncom[str_detect(syncom$FeatureID, "CKEEMGOL"), ]$genus_strain <- "Methylobacterium_OAE516"  # OAE5165 ?
syncom[str_detect(syncom$FeatureID, "MOEDGFBE"), ]$genus_strain <- "Mycobacterium_OAE908"
syncom[str_detect(syncom$FeatureID, "GDBBDLJH"), ]$genus_strain <- "Paenibacillus_OAE614"
syncom[str_detect(syncom$FeatureID, "NBPBNKAL"), ]$genus_strain <- "Rhizobium_OAE497"
syncom[str_detect(syncom$FeatureID, "NBJMGFPL"), ]$genus_strain <- "Rhodococcus_OAS809"
syncom[str_detect(syncom$FeatureID, "NCLMAKNM"), ]$genus_strain <- "Terrabacter_OAE601"

# Remove Terrabacter and Bacillus reads (not part of the SynCom)
syncom <- syncom[syncom$genus_strain != "Terrabacter_OAE601", ]
syncom <- syncom[syncom$genus_strain != "Bacillus_OAE603", ]
syncom <- syncom[, c(1, ncol(syncom), 2:(ncol(syncom)-1))]
```

## Barplot sequencing depth

```{r, fig.width = 25, fig.height = 4}
summary(apply(syncom[, 3:ncol(syncom)], 2, sum))
barplot(apply(syncom[, 3:ncol(syncom)][, str_detect(colnames(syncom[, 3:ncol(syncom)]), "metaG")], 2, sum))
barplot(apply(syncom[, 3:ncol(syncom)][, str_detect(colnames(syncom[, 3:ncol(syncom)]), "metaT")], 2, sum))
barplot(apply(syncom[, 3:ncol(syncom)][, str_detect(colnames(syncom[, 3:ncol(syncom)]), "metaRS")], 2, sum))

apply(syncom[, 3:ncol(syncom)], 2, sum)
```
More reads for Soil + SynCom. good.

## RPKM

Import saf file

```{r}
saf <- read.table("/home/moyne/Moyne_disk/Soil_project/Final_datasets/REAL_SOIL_SAMPLES/Expt_2/WoL_subset/Nov_22/custom_index/Final_18_strains_with_strainN.saf", 
                    sep="\t", 
                    dec=".", 
                    header = FALSE, 
                  col.names = c("GeneID",	"Chr", 	"Start",	"End",	"Strand"),
                    stringsAsFactors = FALSE, 
                    na.strings = "", 
                    check.names = FALSE, 
                    comment.char = "")

saf$GeneID <- gsub("_gene", "", saf$GeneID)

saf$gene_len <- saf$End - saf$Start

saf[order(saf$gene_len), ]

summary(saf$gene_len)
```

```{r}
syncom <- merge(saf[, c(1, ncol(saf))], syncom, by.x = "GeneID", by.y = "FeatureID")
```


# Data Cleaning

```{r}
# Information extracted from ffn file
metadata <- read.delim("/home/moyne/Moyne_disk/Soil_project/Final_datasets/REAL_SOIL_SAMPLES/Expt_2/WoL_subset/Nov_22/custom_index/18_strains_genenames.txt", 
                    sep="\t", 
                    dec=".", 
                    header = FALSE, 
                    stringsAsFactors = FALSE, 
                    na.strings = "", 
                    check.names = FALSE, 
                    comment.char = "")

metadata$gene <- str_sub(metadata$V1, 2, 15)
metadata$gene_name <- str_sub(metadata$V1, 17)

metadata <- metadata[, c("gene", "gene_name")]
```


```{r}
syncom <- merge(metadata, syncom, by.x = "gene", by.y = "GeneID")
```

```{r, eval = FALSE}
syncom[str_detect(syncom$gene_name, "tRNA") &syncom$gene_len<100, ]
syncom[str_detect(syncom$gene_name, "ribosomal RNA"), ]
syncom[str_detect(syncom$gene_name, "transfer-messenger RNA"), ]

#Proportion of metaRS reads for each
sum(syncom[str_detect(syncom$gene_name, "tRNA")&syncom$gene_len<100, str_detect(colnames(syncom), "metaRS")]) / sum(syncom[, str_detect(colnames(syncom), "metaRS")]) #  0%
sum(syncom[str_detect(syncom$gene_name, "S ribosomal RNA"), str_detect(colnames(syncom), "metaRS")]) / sum(syncom[, str_detect(colnames(syncom), "metaRS")])          #  18%
sum(syncom[str_detect(syncom$gene_name, "transfer-messenger RNA"), str_detect(colnames(syncom), "metaRS")]) / sum(syncom[, str_detect(colnames(syncom), "metaRS")])   # 0%    


syncom <- syncom[str_detect(syncom$gene_name, "tRNA", negate = TRUE)&syncom$gene_len>100, ]
syncom <- syncom[str_detect(syncom$gene_name, "tRNA", negate = TRUE)&str_detect(syncom$gene_name, "S ribosomal RNA", negate = TRUE)&str_detect(syncom$gene_name, "transfer-messenger RNA", negate = TRUE), ]

summary(syncom$gene_len)

syncom[order(syncom$gene_len, decreasing = TRUE), ]
```

```{r, fig.width = 25, fig.height = 4}
summary(apply(syncom[, 5:ncol(syncom)], 2, sum))
barplot(apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaG")], 2, sum))
barplot(apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaT")], 2, sum))
barplot(apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaRS")], 2, sum))

apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaG")], 2, sum)
apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaT")], 2, sum)
apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaRS")], 2, sum)
# Soil_SynCom metaRS samples are better overall (more reads aligned to custom database). This is a good sign as we are looking at count tables after alignment to custom genomes

barplot(apply(syncom[, 5:ncol(syncom)], 2, sum), las = 2)
```


## If not present in metaG, etc... (optional)

```{r, eval = TRUE}
# Create separate matrices for each omic
metaG_mat <- syncom[, str_detect(colnames(syncom), "metaG")]
colnames(metaG_mat) <- str_replace(colnames(metaG_mat), "_metaG", "")
metaT_mat <- syncom[, str_detect(colnames(syncom), "metaT")]
colnames(metaT_mat) <- str_replace(colnames(metaT_mat), "_metaT", "")
metaRS_mat <- syncom[, str_detect(colnames(syncom), "metaRS")]
colnames(metaRS_mat) <- str_replace(colnames(metaRS_mat), "_metaRS", "")

# order them the same
metaT_mat <- metaT_mat[, colnames(metaG_mat)]
metaRS_mat <- metaRS_mat[, colnames(metaG_mat)]

# Rename columns
colnames(metaG_mat) <- paste(colnames(metaG_mat), "metaG", sep = "_")
colnames(metaT_mat) <- paste(colnames(metaT_mat), "metaT", sep = "_")
colnames(metaRS_mat) <- paste(colnames(metaRS_mat), "metaRS", sep = "_")

# -------------NICELY DONE HERE

# metaG averages (remove metaT and metaRS if absent of BOTH metaG replicates)
metaG_av <- data.frame(t(metaG_mat))
metaG_av$Community <- str_split(rownames(metaG_av), "_", simplify = TRUE)[, 1]
metaG_av$Time <- str_split(rownames(metaG_av), "_", simplify = TRUE)[, 3]

# Filtering
table(is.na(metaG_mat))
table(metaG_mat == 0)
table(metaT_mat == 0)
table(metaRS_mat == 0)

#metaT_mat[is.na(metaG_mat)] <- 0
#metaT_mat[metaG_mat == 0] <- 0
#metaRS_mat[metaG_mat == 0] <- 0
# metaRS_mat[metaT_mat == 0] <- 0
#metaT_mat[metaRS_mat == 0] <- 0

syncom <- data.frame(cbind(syncom[, 1:4], metaG_mat, metaT_mat, metaRS_mat))
```


## RPKM normalization

```{r}
rpkm_fun <- function(x) x / ( (syncom$gene_len/1000) * (sum(x) / 1000000))
syncom[, 5:ncol(syncom)] <- apply(syncom[, 5:ncol(syncom)], 2, rpkm_fun)    
```

## Barplot sequencing depth after normalization

```{r, fig.width = 9, fig.height = 4}
apply(syncom[, 5:ncol(syncom)], 2, sum)
barplot(apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaG")], 2, sum), las = 2)
barplot(apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaT")], 2, sum), las = 2)
barplot(apply(syncom[, 5:ncol(syncom)][, str_detect(colnames(syncom[, 5:ncol(syncom)]), "metaRS")], 2, sum), las = 2)

barplot(apply(syncom[, 5:ncol(syncom)], 2, sum), las = 2)
```

#--------------------------------

# KEGG Annotation

## Data import (KEGG annotated genomes)

```{r, eval = TRUE}
genes_KEGG1 <- read.table("/home/moyne/Moyne_disk/Soil_project/Final_datasets/18_genomes_KEGG_BlastKOALA.csv", 
                    sep=",", 
                    dec=".", 
                    header = TRUE, 
                    stringsAsFactors = FALSE, 
                    na.strings = "", 
                    check.names = FALSE, 
                    comment.char = "")


dim(genes_KEGG1) # 111949 genes
genes_KEGG1 <- genes_KEGG1[!is.na(genes_KEGG1$Category), ]
nrow(genes_KEGG1) / 111949 * 100 # 23.6% genes annotated

genes_KEGG1[order(genes_KEGG1$gene), ]

genes_KEGG1 <- genes_KEGG1[, c(2:ncol(genes_KEGG1))]

rmarkdown::paged_table(head(genes_KEGG1))

genes_KEGG1[str_detect(genes_KEGG1$pathway, "tryptophan.=>.rebe"), ]
genes_KEGG1[str_detect(genes_KEGG1$kegg, "K00588"), ]

genes_KEGG <- genes_KEGG1
```


## Merge files and aggregate

```{r}
syncom_KEGG <- merge(genes_KEGG, syncom, by.x = "gene", by.y = "gene", all.y = TRUE)
```

# Barplot : stay similar. good

```{r, fig.width=15, fig.height=6}
barplot(apply(syncom[,5:ncol(syncom)], 2, sum), las = 2)
barplot(apply(syncom_KEGG[,9:ncol(syncom_KEGG)], 2, sum), las = 2)
```

# Check top abundances

```{r}
syncom[order(syncom$Soil_SynCom_T3_1_metaG, decreasing = TRUE), ] # rRNA +++
syncom[order(syncom$Soil_SynCom_T3_1_metaT, decreasing = TRUE), ] # rRNA +++
syncom[order(syncom$Soil_SynCom_T3_1_metaRS, decreasing = TRUE), ] # hypothetical proteins and rRNA + genes
```


## Export tables

```{r, eval = FALSE}
write.table(syncom, "/home/moyne/Moyne_disk/Soil_project/Final_datasets/REAL_SOIL_SAMPLES/Expt_2/Custom_index/MIND_Soil_SynCom_Custom_index_Expt2_count_table_RPKM.tsv", sep = "\t", col.names = TRUE, row.names = FALSE)

write.table(syncom_KEGG, "/home/moyne/Moyne_disk/Soil_project/Final_datasets/REAL_SOIL_SAMPLES/Expt_2/Custom_index/MIND_Soil_SynCom_Custom_index_Expt2_count_table_RPKM_KEGG.tsv", sep = "\t", col.names = TRUE, row.names = FALSE)

syncom_KEGG <- syncom_KEGG[!is.na(syncom_KEGG$kegg), ]

write.table(syncom_KEGG, "/home/moyne/Moyne_disk/Soil_project/Final_datasets/REAL_SOIL_SAMPLES/Expt_2/Custom_index/MIND_Soil_SynCom_Custom_index_Expt2_count_table_RPKM_KEGG_noNA.tsv", sep = "\t", col.names = TRUE, row.names = FALSE)
```
